+++ i18n/en.json
@@ -780,6 +780,8 @@
   "admin.permissions.group.send_invites.name": "Add Team Members",
   "admin.permissions.group.teams.description": "Create teams and manage members.",
   "admin.permissions.group.teams.name": "Teams",
+  "admin.permissions.group.teams_team_scope.description": "Manage team members.",
+  "admin.permissions.group.teams_team_scope.name": "Teams",
   "admin.permissions.inherited_from": "Inherited from <a>{name}</a>.",
   "admin.permissions.introBanner": "Permission Schemes set the default permissions for Team Admins, Channel Admins and everyone else. Learn more about permission schemes in our [documentation](!https://about.mattermost.com/default-advanced-permissions).",
   "admin.permissions.loadMoreSchemes": "Load more schemes",
